+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - `DidSelectViewModifier` implemented
 - `getDay` implemented on `Date extension`.
 - `toInt` implemented on `String extension`. 
+- `SelectableViewDataSource` implemented.
 
 # 0.12.0